Use fusion-plugin-target.xcconfig file
Note this is the first reference we're making to a project outside of
the SRCROOT (ie. we now officially require the fusion project to be
checked out alongside this one in order to build successfully; we don't
use submodules, although in this case we probably could consider doing
so if we want to make independent builds as easily as possible).

Signed-off-by: Wincent Colaiuta <win@wincent.com>